Conflicts:
	server/src/main/java/cz/muni/fi/pv243/mustech/util/AppInitializer.java

+++ server/src/main/java/cz/muni/fi/pv243/mustech/service/AbstractGenericService.java
@@ -20,22 +20,36 @@ public abstract class AbstractGenericService<T, R extends EntityRepository<T, Lo
     protected abstract R getRepository();
 
     @Override
-    public void saveOrUpdate(T t) {
+    public void saveOrUpdate(T t)
+    {
+        if(t == null)
+        {
+            throw new IllegalArgumentException("Entity is null.");
+        }
         getRepository().save(t);
     }
 
     @Override
-    public void delete(Long id) {
+    public void delete(Long id)
+    {
+        if(id == null || id <= 0)
+        {
+            throw new IllegalArgumentException("Id must be positive.");
+        }
+
         getRepository().remove(getRepository().findBy(id));
     }
 
     @Override
-    public T findById(Long id) {
+    public T findById(Long id)
+    {
+        if(id == null || id <= 0)
+        {
+            throw new IllegalArgumentException("Id must be positive.");
+        }
         return getRepository().findBy(id);
     }
 
     @Override
-    public List<T> findAll() {
-        return getRepository().findAll();
-    }
+    public List<T> findAll() { return getRepository().findAll(); }
 }
